Add git-link package
This makes it easier to get links to files in git projects. I'll be
using the Github functionality.

One point of frustration for me is getting a permalink to a file. (I use
permalinks whenever the link could be discoverable in the future. This
prevents confusion, since the link will always point to the intended
line.) Setting `git-link-use-commit` to `t` causes the behavior to
default to linking to the specific commit, a permalink.