Add default ticket ID to Jira browse
If the sexp at point could be a ticket id, then default to using it as
the argument for the `browse-ticket` function.

This makes it a lot easier to navigate to a ticket that is already
referenced in a buffer. For example, when looking through history in a
magit buffer, our commit messages will include ticket IDs.