Add dumb-jump for code navigation
I had been trying to use robe, but have been thwarted by memoization
changing the `#source_location` of methods. `dumb-jump`, by using
regexes rather than actually running a ruby process, finds the more
relevant definition of a method.

The default search tool that `dumb-jump` uses in git projects is
`git-grep`. However, this doesn't seem to work for me because I have git
set to recurse on submodules and am getting this error:

```
fatal: --untracked not supported with --recurse-submodules
```

So, in order to avoid that issue, I've installed ripgrep (which is
supposed to be fast) and told dumb-jump to use that.