Add yard-mode
This provides highlighting for yard comments.

I've included an extra section to the hook to change the `paragraph-start`
variable when this mode is enabled. This allows `fill-paragraph` to wrap lines
while considering docs for separate parameters, return value, yields, etc. to be
different paragraphs.

Without this addition (independent of `yard-mode`, which is focused only on
faces), a comment like:

```
@param foo [Integer] some parameter with a nice, descriptive name. So
  descriptive that it takes up multiple lines.
@param bar [Integer] another parameter with a description.
```

would be wrapped as:

```
@param foo [Integer] some parameter with a nice, descriptive name. So
  descriptive that it takes up multiple lines.  @param bar [Integer] another
  parameter with a description.
```