Add indentation handling to YARD paragraph-start
I'm not sure why indentation support is necessary here. From the
description of `paragraph-start`:

> This is matched against the text at the left margin, which is not
necessarily the beginning of the line, so it should never use "^" as an
anchor.  This ensures that the paragraph functions will work equally
well within a region of text indented by a margin setting.

So I would have thought that it would work fine without explicitly
matching the whitespace at the beginning of the line. So more
investigation is needed but, for now, this seems to work on indented
comments as well as top-level ones!