Fix: Use vertico in config
TIL that `:init` is called before the package is loaded, and `:config`
is called after. This was working before since I'd already installed
`vertico` on another machine, but broke when I tried using this config
on a new machine.

Changing from `:init` to `:config` causes the package to be loaded
before trying to use the mode.