Be more permissive with Jira ticket boundaries
People will often write commit messages where the ticket number is only
part of the sexp, like:

```
TICKET-1234-Commit message goes here
```

In these cases the old ticket defaulting wouldn't work, because it
matched on the beginning and end of the string. Now we just assume that
the ticket number begins the string and is in the format
`[A-Z0-9+]-[0-9]+`. I haven't run into any issues with over-aggressive
defaulting, so this should fix up a minor nuisance without increasing
friction!